aoo: allow for variable processing block size
this is needed, for example, if AOO is used inside a VST plugin
+ get rid of unncessary intermediate buffer in sink

+++ aoo/src/time_dll.hpp
@@ -18,7 +18,7 @@ namespace aoo {
 class time_dll {
 public:
     void setup(double sr, int nper, double bandwidth, double t){
-        double tper = nper / sr;
+        double tper = (double)nper / sr;
         nper_ = nper;
         // compute coefficients
         double omega = 2 * M_PI * bandwidth * (nper / sr);
